refactor: 聊天室message監聽邏輯移到useInstantMessage fix: ChatPrivateText修正抓取targetUser
