+++ src/components/ChatInput/ChatInput.jsx
@@ -9,14 +9,13 @@ const ChatInput = ({ onSelfSend, roomId }) => {
 
   const handleSend = (e) => {
     e.preventDefault();
-    if (socket) {
-      const time = new Date()
-
-      console.log('At ChatInput', roomId)
-      socket.emit('client-message', text, time, roomId)
-
-      onSelfSend(text, time)
+    const time = new Date()
+    if (roomId === 4) {
+      socket.emit('client-message', text, time)
     }
+    socket.emit('client-message', text, time, roomId)
+    console.log('At ChatInput', roomId)
+    onSelfSend(text, time)
     setText('')
   }
