+++ src/components/ChatRoom/ChatRoom.jsx
@@ -4,16 +4,17 @@ import ChatInput from "../ChatInput/ChatInput"
 import avatar from "../../assets/icons/avatar.svg"
 import ChatBody from "../ChatBody/ChatBody"
 import { useChat } from '../../context/ChatContext';
-
 import { useAuth } from "../../context/AuthContext"
 import { chatTimeFormat } from "../../apis/data"
-
+import useInstantMessage from "../../hooks/InstantMessageHook"
 
 const ChatRoom = ({ headerContext, roomId }) => {
-  const [message, setMessage] = useState([])
+
   const [historyMessage, setHistoryMessage] = useState([])
   const { user } = useAuth() || {}
   const socket = useChat()
+  const { message, setMessage } = useInstantMessage(roomId)
+
   const handleServerRecord = useCallback((res) => {
     console.log('server-record', res)
     if (res === '尚未聊天過，開始發送訊息吧!') {
@@ -33,18 +34,6 @@ const ChatRoom = ({ headerContext, roomId }) => {
     })
   }, [roomId, user?.id])
 
-  // enter room & leave room
-  useEffect(() => {
-    if (socket?.connected && roomId !== 4) {
-      socket.emit('client-enter-room', roomId);
-      socket.on('server-enter-room', (res) => console.log(res))
-
-      return () => {
-        // socket.emit('client-leave-room', roomId)後端會自動leave
-        socket.off('server-enter-room', (res) => console.log(res))
-      }
-    }
-  }, [roomId])
 
   useEffect(() => {
     if (socket?.connected) {
@@ -52,34 +41,10 @@ const ChatRoom = ({ headerContext, roomId }) => {
       console.log('emit record', roomId)
     }
     return () => {
-      setMessage([]);
       setHistoryMessage([]);
     }
 
   }, [roomId]);
-
-  // 監聽上下線
-  useEffect(() => {
-    const handleServerJoin = (res) => {
-      setMessage((prevState) =>
-      [...prevState, { isChat: false, message: res }]);
-    };
-
-    const handleServerLeave = (res) => {
-      setMessage((prevState) => 
-      [...prevState, { isChat: false, message: res }]);
-    };
-    if (socket && roomId === 4) {
-      socket.on('server-join', handleServerJoin);
-      socket.on('server-leave', handleServerLeave);
-
-    }
-
-    return () => {
-      socket?.off('server-join', handleServerJoin);
-      socket?.off('server-leave', handleServerLeave);
-    };
-  }, [socket]);
 
   // 獨立監聽server-record
   useEffect(() => {
@@ -91,32 +56,6 @@ const ChatRoom = ({ headerContext, roomId }) => {
       socket?.off('server-record', handleServerRecord)
     }
   }, [socket?.connected, roomId])
-
-
-
-  // 獨立監聽server-message
-  useEffect(() => {
-    const handleServerMessage = (res) => {
-      console.log('server-message', res)
-      if (Number(roomId) !== Number(res.room)) return
-
-      const other = { text: res.message, time: chatTimeFormat(res.timestamp), avatar: res.user.avatar, isOwner: res.user.id === user.id }
-      setMessage((preState) => [...preState, { isChat: true, message: other }])
-
-    }
-
-    if (socket) {
-      console.log('im lisening')
-      socket.on('server-message', handleServerMessage);
-    }
-
-    return () => {
-      console.log('not lisening')
-      socket?.off('server-message', handleServerMessage);
-
-    }
-  }, [socket, roomId])
-
 
   // 接收來自ChatInput的props
   const handleSelfSend = (text, time) => {
@@ -135,6 +74,7 @@ const ChatRoom = ({ headerContext, roomId }) => {
     </>
   )
 }
+
 export default ChatRoom
