+++ src/context/ChatUserContext.jsx
@@ -1,13 +1,10 @@
 
 import { useContext, createContext, useState } from 'react';
-import { useAuth } from './AuthContext';
 export const ChatUserContext = createContext();
 export const useChatUser = () => useContext(ChatUserContext);
 
 export function ChatUserProvider({ children }) {
   const [chatUser, setChatUser] = useState({});
-
-
 
   return (
     <ChatUserContext.Provider value={{ chatUser, setChatUser }}>
